Hugely improve collision detection.
This code does more than one thing, but teasing it back apart into
smaller bits just isn't worth the effort -- so here is a list:

First, add a stupid ton of logging using the debug-logging plumbing
that was introduced earlier. When the global debug symbol is on, each
piece has its movements logged and most of the collision detection and
piece-reset process is traced out. This was very useful in figuring out
what was wrong and how to fix it, and as there are more problems to
solve, it remains.

Second, add a tracking mechanism to determine what pieces have been
moved in each step, and therefore what pieces actually need to be reset
to their last_x/last_y. This prevents accidentally resetting a piece
that hasn't been moved, prevents not resetting pieces that have been
moved, and also prevents double-resets in case halt() gets called
multiple times, which I'm not a good enough programmer to guarantee
will never happen. :P

Third, fix a big problem where the piece that encounters a collision
doesn't register a movement, but keeps new x/y coords. This was a source
of shapes "coming apart" after collision, which was likely the most
visible problem in the program. The new code restores last_x/last_y in-
line before calling halt(). A different way would be to mess with the
halt/track_movement/move order, but that felt risky to do in the midst
of so much moving code. Will look back into it later.

Fourth (I don't know if I'll keep this one) it pushes the x/y vals to
be integers in an attempt to make pieces "stick" better after collision,
which is not the answer. If a piece is moving at high speed and its next
move would cause it to collide, it stops X pixels away from its barrier,
and just rounds the pixel out. It should align to the next barrier --
I will try to fix that next, likely.