Fix bounds issues.
Pieces now stay inside the pretty box.

The core of this issue was not adding the size of a tile to the highest
x and y values, which led to a vast amount of experimentation with the
offsets, as I thought a miscalculation of bounds and offsets was the
root cause. It wasn't.