...and, it's now tetris.
You can now clear lines.

This is implemented just inline, as a total mess, just to get the
basic thing to work. And it does! I played it for about 10 minutes.

You can clear 1, 2, 3, or 4 lines at a time now, and pieces drop as
you'd expect them to. There's unfortunately not yet any scorekeeping
or level progression or pause functionality. But it's totally tetris.