Clamp pieces into the grid at spawn.
Some conditions could lead to pieces spawning above the grid and not
getting a chance to drop at all, which then led to attempts to place
them on pieces of the grid that don't exist! The game was gonna have
a bad time. So they get .. the clamps!

Also deletes a console.log that was spammy and no longer useful.