Implement pause.
This has some hackey behaviour in it -- for example, when the level
is first loaded, the "last paused" value is set in-line, to make sure
that the user hitting 'enter' to start the game can't accidentally
trigger pause right off the bat. I need to deal with timing issues like
that in a uniform and smart way. Problem for later!

Also, I am really just pausing the game.update and player.update calls
manually. The game should have a 'paused' state in the engine that I
initiate wherein no updates are triggered, or, something. The entities
that make up the piece are only playing nice here because they sit on
the player, and the player can't move. But it works for tetris.