Add score+row tracking and persistence.
This is just super basic crap storing it in a cookie in a terrible
way, but it means there's at least the shadow of a reason to play.

I've version-namespaced the cookie as "v0" so that I can hopefully
do better in a graceful way down the road. Scores are +200/line cleared
and + random 3-9 per block placed. I will likely add bonuses for two-
at-a-time and quadruple clears later, and I really want to add special
score bonuses for clearing lines above a certain height or clearing
the whole course down to a certain point after it has been up high.

Also, I've got to implement a level progression! And sounds and music!