Add audio manager and empty_dict util.
These will be added upstream to creek, where I really ought to start
using versions to track things, perhaps!

This is very poorly done at the moment. The entire game will have to
sit and wait for all images and all sounds to load before it moves
forward, which in my current testing means sitting on a blank screen for
10+ seconds -- it also operates on a silly sort of fallback behaviour
where if a sound clip is not loaded, the audio manager will ask the
resource manager for it, and this will continue every time it's
requested, but there's no great way to tell from the manager level that
a request to say, play a file has failed, and that it should be retried.

I see an opportunity to re-jig the loading mechanism entirely here,
but it's a much larger change and will require work on the creek demos
too -- it will impact images as well as sound, registering all of what
will be loaded before it is loaded and then letting the engine move
forward with its work while things load.

There is also something funny going on with the 'canplaythrough' event,
where it is registering every time a piece of audio finishes playing --
I either need to de-register my events from that trigger upon the first
execution of it, or I need to find a different trigger that's explicitly
just for loading, as opposed to one that happens when a track loads OR
finishes playing.

Aside from those limitations, and the odd approach of nearly duplicating
all of the code here on two layers (it's to make moving to/supporting
the WebAudio API easier, which I think I will want to do at some point),
this code does a fair job! It can play/pause/stop/mute/set volume on
individual tracks which do not interfere with each other, and it has
operations that apply to all tracks simultaneously.

It can also pause all currently playing tracks and then resume just
those tracks. An interesting/odd piece to this is that repeated calls
to "pause_all" will aggregate tracks to be unpaused, which felt handly,
but might prove awkward long term. We'll see!

Oh also, add an "empty_dict" method to the utilities, because {} does
not evaluate as false, and checking emptiness of dicts is a surprisingly
involved little operation to sprinkle everywhere.