Introduce a Manager Manager.
The EntityManager had become a Manager of Managers in addition to its
regular job, which meant it was going to have a bad time. In fact it
was already having a pretty bad one. So, here's a small novel of a
commit message explaining why it was time to fix this, what was done,
and why it was done that particular way.

The following things added up over time, and finally overcame both my
concern that a specific manager-of-managers in everyone's pocket would
cause a situation not unlike recklessly used globals, and my fear of the
amount of work it would take to switch all of the code I've got over:

  - the mental weight of tracking what Managers were passed where
  - the code weight of adding them to multiple places for any new need
  - the code weight just of having N arguments passed through M layers
  - the mental/maintenance weight of shifting Manager init orders around
    (this happened a stupid amount!)

In practice, I have mostly been treating Managers as though they could
be available to anyone in most circumstances, so I don't think this new
ease-of-access is going to destroy my oh-so-pristine code, but even if
it does, that mess currently seems like a better one than the mess I
have.

I'm about to implement some changes to the way loading resources works,
which may require the ResourceManager or RenderManager or AudioManager
to have access to Managers they don't currently use. I feel a deep sort
of foreboding sense that the job of fixing the EntityManager's
superpowered-ness will be a lot harder after I fix the loading stuff,
and also happen to optimistically believe that fixing the loading stuff
will be a lot easier if I've fixed up this conceptual baggage first. So
there's no time like the present!

Onto the meat: what does it do? what's changed?

The Manager Manager registers all of the other Managers, and gives
access to them. Every Manager has a reference to the Manager Manager,
and thereby can get access to any other Manager. Which by the way, are
called 'Managers' because they manage access to canonical versions of
data in your Creek app, which are not always easy to have a reference to
across closures and callbacks and other grim realities of javascript.

Previously, each Manager was declared as an immediately-eval'd function
that encapsulated the scope of the data it cared about, and returned a
function that acted as a constructor. That constructor function is what
was assigned to each Manager. It did three things: take in the arguments
needed for construction, call a special private 'init' function and pass
those arguments to it, and then return a dictionary of the Manager's
exposed methods, which it had access to via its encapsulated scope, and
which all operated on (or gave access to the canonical version of) the
other data under the Manager's watch, which also resided in that scope.

The Manager Manager could have merely stored references to the already-
instantiated other Managers, which would have allowed this system to
be preserved nearly intact, but that would sacrifice most of the code-
weight savings of having a Manager Manager at all. Having one is nice,
but not that nice. So a pretty major change was undertaken here, as is
evidenced by the diff and by the vast sea of text you're swimming in.

The Manager Manager is now responsible for the initialization of all
other Managers (which may actually not be necessary; I'll examine
that some time later[1]), but not their declaration or instantiation.

The various Managers are created and then passed to it to be managed.
That requires giving access to the earlier 'init' method to outside
callers, (namely the Manager Manager) and allows us to now pass *just*
the Manager Manager into each Manager, rather than a list of all of
their dependencies as arguments to first the constructor, and then
again to the private init function, and then pretty much a third time
as they are each written to encapsulated-scope-references held by the
Manager. Now the init methods just ask the Manager Manager for whatever
they need!

This can (and will) be taken further, so that fewer or no references to
Managers aside from the Manager Manager are stored in individual
Managers' encapsulated scopes. They can just ask for whatever they need
at the time of use. I hope that this will have a noticeable positive
impact on the mental and screen weight of the code.

A pitfall is that this would brutally messes up the initialization
order of the Managers, which up until now was carefully orchestrated to
ensure that dependencies were prepared in time to be used. The Manager
Manager is also quite helpful here: it stores whether a given Manager
has been initialized and if a request is made for an uninitialized
Manager, the Manager Manager will initialize it first, then return it.

"But there's a risk of cycles!" -- correct. Manager A's init method
could ask for Manager B's init method which could then ask for Manager
A's init method, and so on. To avoid this, the Manager Manager marks
Managers as initialized just *before* their init calls, rather than
just *after*. The tradeoff is that instead of easy to spot showstopping
loops, we may get unprepared Managers being used. That could be harder
to notice and track down, but in practice most Managers only need a
reference to other Managers during init, not to fully-prepped data.

Cases where a Manager *does* need fully-initialized data from another
Manager during init should be rare enough... to manage. Rather than
stop the world and frustrate the developer (i.e. me) by requiring
careful orchestration to avoid cycles, we'll be chill about it and give
said developer more than enough rope to hang themselves with. Hopefully
it's as rare an occasion as it seems.

And that's about it! There are a few small fixes in this change that
would just be a bit of a bother to pull out, like adding 'let' to a
couple of declarations that were accidentally globally scoped, adding
some more Manager init messages, etc. And obviously the required bits
of work in config.js and app.js are also thrown in here -- they should
be self-explanatory.

Upcoming changes are to make more use of the Manager Manager, mostly
around reducing the Entity Manager's double duty as a pseudo manager
manager, and letting it just make entities great again.

Lastly (aside from the two postscripts, jesus), "Manager Manager" and
"FooManager" are terrible names and the word "manager" occurs in this
document 83 times. That's nuts! I'm going to think about how to fix it,
likely by using 'Repo' or 'Service' or something in some places.

1 - I'm currently explicitly calling init on every Manager, but I don't
think it's necessary in practice. My intuition is that, if a Manager is
needed, it will be used, and therefore it will be requested, and in the
current structure of the code, it then must be initialized. Any Manager
that doesn't get initialized just isn't being used by the app. I think.
I've got to do some testing to verify this, but soon the ManagerManager
may not even call the various inits. Trust the graph.

2 - This commit message will be more or less copied over to the actual
creek project, probably with some revisions as I will likely have
done the next few pieces of work by then to reduce integration burden.