Fix non-unique timestamp ids.
The drop in performance.now() resolution introduced as a mitigation
for Meltdown/Spectre attacks led to my timestamp id function spitting
out duplicate ids with absurd frequency. Which basically made the game
go insane.

It took a lot of digging to realize that this was actually the issue, as
opposed to some silly issue with badly scoped variables. The next commit
is dropping 'lets' _everywhere_ and some other stuff designed to ferret
out what I thought the problem would be.

The silly bad solution to the drop in resolution is to add
Math.rand()*100 to the output of performance.now(), which *mostly*
works, but still does have rare duplication. Since this was only
impacting the piece generation, I then stuck the piece index directly
into its id along with the timestamp, and threw a uniqueness checker
in afterward.

I'll have to rethink the timestamp id idea.