Edge case bugfixes.
In one of these, particularly in the duplicate-id case, phantom pieces
could end up on the board that wouldn't be found when clearing lines
later, so having a check on having actually found the entity before
moving it is wise.

In the other, a rare edge-case where happening to hit rotate at the
exact moment a piece has transitioned to a static piece and been nulled
out would cause rotate to be called on a null shape. That would be bad!
So a guard is added to ensure that player.shape is non-null before
issuing the rotate.