Fix another duplicate timestamp issue.
Timestamps in the game were being generated for pieces to look like
this:

piece-3455_000

with just 0, 1, 2, or 3 as the leading number, and the reset being
the last few characters of the timestamp, with its lower precision
since the spectre/meltdown mitigations.

In practice, this means that the last 2 digits are rarely if ever
set to anything but 0, and we only have ~3-4 digits of uniqueness at
best. Unlike the first duplicate bug which came down to the odds of
generating two pieces with the exact same timestamp, this newly found
bug comes down to two pieces having the same information in the bit of
the timestamp that we bother to keep -- over a reasonable playtime,
this becomes quite likely! this is a little bit similar to the
birthday paradox.

So my solution is to just to keep 3-4 more digits, which doesn't "fix"
the problem, but makes it very likely to be fine even on hours and
hours and hours of play. They will now look like this:

piece-1-440531_900

With the piece-index clearly called out and 3 extra digits of info
included. At some point I'm going to have to look into a better scheme
for id generation, that guarantees uniqueness. I hate to jump straight
to GUIDs or UUIDs, but that might be the most sane thing!