Again, improve collisions -- use better AABB.
I hear that this algorithm is a much more common one, and I'm using
it in the place of the circle collision detection algorithm I'd had
previously underpinning the main entity collision code. This one thing
made a lot of other stuff suddenly work, which upsets me, because I...
don't really understand the underlying reason the things were broken.

Anyway, it enabled some good stuff that'll show up in the next few
commits. A sideshow in here is changing the fine-grained collision
code a bit too, in ways that are confusing. I'm on the edge of going
back and ripping that code out, so I don't want to get deep into the
(probably flawed) rationale for the small changes -- it was an attempt
to make boundary-touching to not count as collisions.