Add shape-sliding.
This took a long time to get to work -- not the sliding itself, which
took maybe 15 minutes -- but allowing a player to slide a block of say
2-height into a gap that was 2-high. Unwanted y-collisions and to my
surprise, x-collisions were cropping up and required the previous
commit's work to use a better fundamental collider for the model.

The pieces transition from falling to sliding, and once in the sliding
state they will automatically transition to static after a set period
of time. If they move out of y-collision while in sliding, they're set
back to falling -- this was a big cause of trouble, because (with the
way collisions work in my engine) I had to tell them to "keep dropping"
while sliding, which can cause misalignment and some of those unwanted
y-collisions I mentioned earlier.

But once the collider was improved, it works! You can slide pieces all
over the place and they fit as desired. Rotations even work in some
very tight spots -- it's a beautiful thing.

There's one known bug that I've witnessed once -- I was able to rotate
one piece into another with this code, but I haven't been able to repro
it. I'm not certain if it was a wild edge case in extant code or is
somehow related to the new collider.