Add progression and random modes, and speeds.
These aren't accessible via the UI yet -- that'll come soon. Pretty
hackily just tossed the data into the player object for now. Just
increase the "level" or randomize it every progression_length number
of cleared rows, which is a tool to let the player choose the length
of their game -- I'll often want to play a much longer game that's
still getting harder, but someone may want to play a game that ramps
up pretty quickly to force them to work hard.

This is likely to see a lot of tuning, but I've known I needed to add
it for .. over a month, haha. Better to just ship it.