Update Creek engine.
Creek is being used for a different project, and occasionally needs
to be modified or improved for that work -- part of due diligence in
that effort is to bring all of the work over here and test its effect.

I found two bugs in Creek's improvements (a dependency cycle and that
a locality optimization for collisions caused issues colliding with big
map boundary objects, more info in Creek commit logs from today's date),
and found a few small bugs in Tetris too -- more or less just odd things
around using "fullscreen" mode, setting a renderable size, and moving
the camera around.